fix on the set amount error
